Add LoadState and GetState to Tape and remove Wrap
The idea here is that we don't want to expose our internals
except thru the use of the GetState method which in effect clones
our state without exposing any of our internal implementaiton details.

Load state allows us to create a new Tape in a prearrange non-default
state.

The use of these two methods allows us to unit test methods like
MoveForward and MoveBackward (and all the others) without having
to resort to side effects. We can call MoveForward on a tape in a known
state, call GetState and confirm that the position changed.

Unit tests updated to pass. Need to write more unit tests to get
100% code coverage on Tape class.