Add support for input.
Added a input of type text which can accept one character at a time. It
is disabled by default. Only when an input instruction is being executed
is it enabled. Immediately after the keyUp event the input is disabled.

When an input instruction is being executed the fetch/execute cycle is
terminated and the program does nothing until the user types a character.
The fetch/execute cycle is then resumed where it left off.