Convert the displaying of output to MVVM
Removed the code that required passing in the output node to the
brainmess object and then to context. Instead a callback function is
passed in that then is invoked whenever the output string changes.

A simple view model with just the output string was added. And
then a simple view for the ouput (bound to the view model) was add
to the html page.